refactor complete on base level. Always room for more, but much better than the first go.
