--- refactor-demo/data/thank-you-codeFellows/page-1.json
@@ -1,96 +1,0 @@
-{
-  "data": {
-    "stats": [
-      {
-        "statOne": {
-          "stat": "In-field Job Placement",
-          "value": "93%"
-        },
-        "statTwo": {
-          "stat": "Unique Hiring Companies",
-          "value": "800+"
-        },
-        "statThree": {
-          "stat": "In Scholarships Awarded ",
-          "value": "$3.5M+"
-        },
-        "statFour": {
-          "stat": "Powered by Schools",
-          "value": "16"
-        }
-      }
-    ],
-    "upComingCourses": [
-      {
-        "upcomingCourse": {
-          "course": "Code 102: Intro to Software Development",
-          "calendarTime": "Mar 1, 2021 - Mar 31, 2021",
-          "typeOfCourse": "Self Paced Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Code 201: Foundations of Software Development",
-          "calendarTime": "Mar 8, 2021 - Apr 2, 2021",
-          "typeOfCourse": "Day Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Code 101: Explore Software Development",
-          "calendarTime": "Mar 13, 2021",
-          "typeOfCourse": "Self Paced Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Ops 101: Explore Ethical Hacking & Cybersecurity Careers",
-          "calendarTime": "Mar 13, 2021",
-          "typeOfCourse": "Self Paced Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Code 301: Intermediate Software Development",
-          "calendarTime": "Mar 15, 2021 - Apr 9, 2021",
-          "typeOfCourse": "Day Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Code 301: Intermediate Software Development",
-          "calendarTime": "Mar 20, 2021 - May 20, 2021",
-          "typeOfCourse": "Day Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Code 102: Intro to Software Development",
-          "calendarTime": "Mar 22, 2021 - Mar 26, 2021",
-          "typeOfCourse": "Day Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Code 401: Advanced Software Development in Java",
-          "calendarTime": "Mar 22, 2021 - May 28, 2021",
-          "typeOfCourse": "Day Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Code 101: Explore Software Development",
-          "calendarTime": "Mar 28, 2021",
-          "typeOfCourse": "Self Paced Course"
-        }
-      },
-      {
-        "upcomingCourse": {
-          "course": "Ops 101: Explore Ethical Hacking & Cybersecurity Careers",
-          "calendarTime": "Mar 28, 2021",
-          "typeOfCourse": "Self Paced Course"
-        }
-      }
-    ]
-  }
-}
